// email-designer-app/src/app/email-designer/models/block-style.interface.ts
export interface BlockStyle {
  padding?: number;
  backgroundColor?: string;
  color?: string;
  fontSize?: number;
  width?: number;
  textAlign?: 'left' | 'center' | 'right' | 'justify';
  borderType?: 'transparent' | 'light' | 'solid';
  borderWidth?: number;
  borderColor?: string;
  borderRadius?: number;
}